fix: generation of __typename
